refactor: prepare for view-only mode
